# Optimize gradle build time