@juhoautio 2019-01-25
The quickest build in master build history's first page is `4 min 45 sec`.

This PR built on Travis `4 min 57 sec`.

On the other hand azkaban builds on Tarvis may some times take even 7 minutes.

Need more data to say anything sure about does this have an impact on build times on Travis on not.

If it's at least not making things worse, I'd like to have this change to optimize builds on my personal machine – which hopefully helps others too.