@juhoautio 2019-01-28
@HappyRay @jamiesjc anyone to test on your machine if you can see similar speed-up with this change?