@burgerkingeater 2019-02-12
according to https://github.com/azkaban/azkaban/pull/1231, enabling maxParallelForks should reduce the test running time, @juhoautio  have you compared setup without maxParallelForks specified?