@juhoautio 2019-02-12
@burgerkingeater yeah that's the theory, but the reality in practice seems different.

These are just guesses of course:

> Maybe the tests interfere with each other.. Maybe the additional forking is just expensive because every test needs to load some heavy classes / static members again..

But:

> Whatever the reason is, there's no point in having parallelism on if makes the total build time longer.

If I entirely remove `maxParallelForks` the execution time is on the same level as with `maxParallelForks = 1`.

Could you also test it on your machine locally?