@juhoautio 2019-02-21
@burgerkingeater have you been able to check what your local build time is and does this setting affect it?