@burgerkingeater 2019-02-27
@juhoautio  i did the experiment, it's strange with `maxParallelForks = 1`, `./gradlew cleanTest test` failed for some tests and it failed in 2m 37s,  and `maxParallelForks = 12` succeeded in  2m 10s.