@juhoautio 2019-02-27
@burgerkingeater tests shouldn't fail because of the maxParallelForks setting, at least when the value is 1. So maybe it was just some random error. Did you try only once? Maybe test once more with different settings?

If the conclusion is that at least maxParallelForks isn't making tests faster on any setup, would you be open to setting it to 1 in the repo?