@juhoautio 2019-03-04
@burgerkingeater are we able to conclude that this change can only make test run faster? In some environments it doesn’t, but it also doesn’t slow it down. Right? Please, this would be helpful for me at least, probably also for at least some other developers also. Now about a month old pr, so would be nice to reach a conclusion.