@burgerkingeater 2019-03-05
@juhoautio sorry for the delayed response. It's strange that the test run time is increased with https://github.com/azkaban/azkaban/pull/1231. I also observed the same result https://github.com/azkaban/azkaban/pull/1231#issuecomment-469506796. Can you summarize our experiment and refer this PR to https://github.com/azkaban/azkaban/pull/1231, then I will merge it, thanks.