@ypadron-in 2019-03-27
> @HappyRay @jamiesjc anyone to test on your machine if you can see similar speed-up with this change?

Ran several times `./gradlew cleanTest test` with `maxParallelForks = 1` in a MacOS' 10.14 terminal. This change reduced running time from 1m 15s (with maxParallelForks=12) to 28s overall.

Should we consider set `maxParallelForks = 1` to the TRAVIS configuration inside the build.gradle file as well? Currently is set `maxParallelForks = 2`